[Feature #158811369v] Fix server side tests
- Write more tests to cover more test cases
- Refactor old tests based on new implemenation of code logic
- Refactor code to pass tests

[Finishes #158811369]
