+++ package.json
@@ -102,6 +102,7 @@
     "jquery": "^3.2.1",
     "jsonwebtoken": "^8.1.0",
     "jwt-decode": "^2.2.0",
+    "lodash": "^4.17.10",
     "morgan": "^1.9.0",
     "multer": "^1.3.0",
     "node-sass": "^4.7.2",
