+++ server/middleware/authenticate.js
@@ -1,6 +1,8 @@
 import dotenv from 'dotenv';
 import jwt from 'jsonwebtoken';
+import db from '../models';
 
+const { User } = db;
 dotenv.config();
 
 /**
@@ -16,10 +18,16 @@ const authenticate = (req, res, next) => {
   const token =
     req.body.token ||
     req.headers['x-access-token'] ||
-    (req.headers.Authorization && req.headers.Authorization.slice(7)) ||
+    (req.headers.authorization && req.headers.authorization.slice(7)) ||
     req.query.token;
 
   if (token) {
+
+    // if user is not in the db, throw err and delete token
+
+
+    // else decode token
+
     jwt.verify(token, process.env.SECRET, (error, decoded) => {
       if (error || !decoded) {
         return res.status(401).json({
@@ -27,8 +35,23 @@ const authenticate = (req, res, next) => {
           message: error.message || 'token issues',
         });
       }
+      // console.log(decoded);
       req.user = decoded;
-      next();
+      const { userId } = decoded;
+
+      User.findById(userId).then(user => {
+        if(!user) {
+          return res.status(404).json({
+            success: false,
+            message: 'User does not exist!'
+          })
+        }
+        next();
+      }).catch({
+        success: false,
+        message: 'Something went wrong, could not query user'
+      });
+
     });
   } else {
     return res.status(401).json({
